+++ src/controllers/User.Controller.js
@@ -4,7 +4,7 @@ import jwt from 'jsonwebtoken';
 import { sendEmail } from '../helperFunctions/HelperFunctions.js';
 import { userRegisterValidator, userResetValidator } from '../validators/User.Validator.js';
 import UserModel from '../models/User.Model.js';
-import { handleValidationError, handleUserExists, handleUserNotFound, tryCatchWrapper, handleWrongCredentials, handleInvalidUser } from '../factory/Factory.js';
+import { handleValidationError, handleMissingParamsError, handleUserExists, handleUserNotFound, tryCatchWrapper, handleWrongCredentials, handleInvalidUser } from '../factory/Factory.js';
 
 
 export const createUser = async (req, res) => {
@@ -14,22 +14,16 @@ export const createUser = async (req, res) => {
             handleValidationError(error, res);
             return;
         }
-
         const { fullName, email, userName, password, phoneNo, course, yearOfStudy, techStack } = req.body;
-
         const existingUser = await UserModel.findOne({ $or: [{ userName: userName }, { email: email }] });
-
         if (existingUser) {
             handleUserExists(res);
             return;
         }
-
         if (await UserModel.create({ fullName, email, userName, password, phoneNo, course, yearOfStudy, techStack, userType: 'user' })) {
             res.status(201).json(req.body);
         }
-
     };
-
     tryCatchWrapper(handler, req, res);
 };
 
@@ -44,6 +38,10 @@ export const getUsers = async (req, res) => {
 export const getUser = async (req, res) => {
     const handler = async (req, res) => {
         const { id } = req.params;
+        if (!id) {
+            handleMissingParamsError(res);
+            return;
+        }
         const user = await UserModel.findById(id);
         if (condition) {
             const { password, ...userWithoutPassword } = user._doc;
@@ -56,7 +54,15 @@ export const getUser = async (req, res) => {
 export const updateUser = async (req, res) => {
     const handler = async (req, res) => {
         const { id } = req.params;
+        if (!id) {
+            handleMissingParamsError(res);
+            return;
+        }
         const { fullName, userName, email, password, course, yearOfStudy, techStack } = req.body;
+        if (!fullName && !userName && !email && !password && !course && !yearOfStudy && !techStack) {
+            handleValidationError({ details: [{ message: "At least one property must be updated" }] }, res);
+            return;
+        }
         const user = await UserModel.findById(id);
         if (!user) {
             handleUserNotFound(res);
@@ -79,15 +85,23 @@ export const updateUser = async (req, res) => {
 export const deleteUser = async (req, res) => {
     const handler = async (req, res) => {
         const { id } = req.params;
+        if (!id) {
+            handleMissingParamsError(res);
+            return;
+        }
         const user = await UserModel.findByIdAndDelete(id);
         user ? res.status(200).json({ message: "User deleted successfully" }) : handleUserNotFound(res);
     };
-
     tryCatchWrapper(handler, req, res);
 };
 
 export const loginUser = async (req, res) => {
     const handler = async (req, res) => {
+        const { error } = userLoginValidator(req.body);
+        if (error) {
+            handleValidationError(error, res);
+            return;
+        }
         const { userName, password } = req.body;
         const user = await UserModel.findOne({ userName: userName, password: password });
         if (user) {
@@ -101,7 +115,6 @@ export const loginUser = async (req, res) => {
             res.status(200).json({ token: token, user: userWithoutPassword });
         } else handleWrongCredentials(res);
     };
-
     tryCatchWrapper(handler, req, res);
 }
